+++ RELEASE_NOTES.md
@@ -39,7 +39,9 @@ KeyboardKit Pro adds new settings screens, like `KeyboardApp.SettingsScreen` and
 * `FeedbackService` is a new service type that can trigger audio and haptic feedback.
 * `Image` has many new utility functions to make it easier to resolve keyboard icons.
 * `Keyboard.BottomRow` is no longer a Pro feature, but is available in the core library.
+* `Keyboard.Services` has a new `frequentEmojiProvider` and removes the static instance.
 * `Keyboard.StorageValue` is a new type that is used to persist codable types in storage.
+* `KeyboardAction.StandardHandler` uses an emoji provider instead of an injected handler.
 * `KeyboardContext` has new settings that replace the ``KeyboardSettings`` settings properties.
 * `KeyboardContext` has a new ``addedLocales`` settings that can be used to "activate" certain locales.
 * `KeyboardContext` now persists the `keyboardLocaleIdentifier` and restores the locale on next launch.